Graphing and Serial Fixed
The serial communication with the Arduino is now functional, and is
able to be extended for later use.  Graphing is now supported for two
parameters, including warning bars which will flash when passed.
Bit-Fields have also be implemented for tracking the status of both the
serial interface, as well as the box’s error states.