Processing Functional
Processing code has been updated to record to .csv properly.  The
naming functionality has also been implemented.  The GUI is almost
complete, all that is needed is numerical representations for the
temperature/humidity, and text to describe the state of the Serial
connection.  Before full functionality, all warnings must be
implemented (such as the social time warning).

The Arduino code has been modified to include analog reading of
temperature and humidity for box 1.  The door-close signal must be
implemented, as well as the code for all boxes.